Replace CacheControl dependency with usecache arg
Remove dependency on the CacheControl library since it doesn't really
seem necessary to add an extra dependency for caching when, ideally, the
program will only be run one or two times each use. Instead added a
command line argument, `usecache`, that will cause the program to reuse
any GeoNames files that have already been downloaded (but still download
any that haven't been downloaded yet). The purpose of this is to avoid
having to wait for the program to download the files every time if it is
being run repeatedly to debug, etc.