Update scraper for 2021
Several teams had to be manually looked up and added to geo_cache. A new
special fix for certain locations in Taiwan was implemented. For some
teams, it seems that administrative divisions like "Taipei" are instead
listed as "Taipei Special Municipality", which causes the city not to be
found. For the fix, any administrative division name in Taiwan that
comes from TBA API ending in Municipality or Special Municipality
automatically has this removed from the name.